moved calc package outside of texas/ folder
